chore(assets): rename asset dirs
`JS` -> `SCRIPTS`
`CSS` -> `STYLES`
