--- assets/css/welcome.scss
@@ -1,20 +1,0 @@
-
-$images-dir: '../images/';
-
-.sticker-container {
-  // background: repeat url('#{$images-dir}generic-tranparent-background.jpg');
-  cursor: pointer;
-
-  img {
-    -webkit-filter: drop-shadow(0 0 0 #222);
-    filter: drop-shadow(0 0 0 #222);
-    transition: filter 0.1s ease-in-out, -webkit-filter 0.1s ease-in-out;
-    max-width: 100%;
-  }
-
-  &:focus img,
-  &:hover img {
-    -webkit-filter: drop-shadow(0 0 5px #222);
-    filter: drop-shadow(0 0 5px #222);
-  }
-}
