--- assets/js/password/forgot-request.ts
@@ -1,1 +1,0 @@
-import 'CSS/password/forgot-request.scss';
