--- assets/js/welcome.ts
@@ -1,1 +1,0 @@
-import 'CSS/welcome.scss';
