+++ assets/scripts/ada/EmulateButton.ts
