+++ assets/scripts/app.ts
@@ -1,10 +1,10 @@
 import $ from 'jquery';
-import 'CSS/app.scss';
+import 'STYLES/app.scss';
 import M from 'materialize-css';
 import 'materialize-css/dist/css/materialize.css';
 import '@fortawesome/fontawesome-free/css/all.min.css';
 
-import MakeClickable from 'JS/ada/EmulateButton';
+import MakeClickable from 'SCRIPTS/ada/EmulateButton';
 
 $(() => {
   $('.is-button').each((_, element) => {
