+++ assets/scripts/lib/loading-spinner.ts
