+++ assets/scripts/notes/code-mirror-assets.ts
