+++ assets/scripts/notes/full-sample.md
