+++ assets/scripts/password/forgot-request.ts
@@ -1,0 +1,1 @@
+import 'STYLES/password/forgot-request.scss';
