+++ assets/scripts/types/Generic.ts
