+++ assets/scripts/users/login.ts
@@ -1,2 +1,2 @@
 // Importing this file, will auto-show the Consent popup
-import 'JS/lib/cookie';
+import 'SCRIPTS/lib/cookie';
