+++ assets/scripts/welcome.ts
@@ -1,0 +1,1 @@
+import 'STYLES/welcome.scss';
