+++ assets/styles/app.scss
@@ -1,6 +1,7 @@
 // @import 'materialize-override.scss';
-@import 'CSS/app-variables.scss';
-@import 'CSS/cookie-consent-overrides.scss';
+@import 'STYLES/app-variables.scss';
+@import 'STYLES/cookie-consent-overrides.scss';
+@import 'STYLES/materialize-override.scss';
 
 html {
   background-color: var(--page-background);
@@ -178,4 +179,10 @@ input {
       height: 5em;
     }
   }
+}
+
+a[target="_blank"]::after {
+  content: " \f35d";
+  font-family: "Font Awesome 5 Free";
+  font-weight: 900;
 }
