+++ assets/styles/notes.scss
