+++ assets/styles/product/admin.css
