+++ assets/styles/welcome.scss
@@ -1,0 +1,54 @@
+.stylish-code {
+  border: solid 1px #d79921;
+  border-radius: 0.25em;
+  background: #00000088;
+  color: #83a598;
+}
+
+main {
+  color: cyan;
+}
+
+h1 {
+  font-size: 2.5em;
+}
+
+h2 {
+  font-size: 2em;
+}
+
+hr {
+  border-top: 1px solid #ff00ff;
+  border-bottom: none;
+}
+
+.fun-list {
+  list-style: none;
+  padding: 0;
+  margin: 0 0 3rem;
+
+  li {
+    padding-left: 1rem;
+    text-indent: -0.7rem;
+  }
+
+  li::before {
+    font-family: "Segoe UI Emoji", "Apple Color Emoji", "Noto Color Emoji";
+  }
+
+  li:nth-child(4n + 1)::before {
+    content: "🕹 ";
+  }
+
+  li:nth-child(4n + 2)::before {
+    content: "📼 ";
+  }
+
+  li:nth-child(4n + 3)::before {
+    content: "👾 ";
+  }
+
+  li:nth-child(4n)::before {
+    content: "👻 ";
+  }
+}
