+++ tsconfig.json
@@ -12,8 +12,8 @@
         "baseUrl": "./",
         "lib": ["esnext", "webworker"],
         "paths": {
-            "JS/*": ["./assets/js/*"],
-            "CSS/*": ["./assets/css/*"],
+            "SCRIPTS/*": ["./assets/scripts/*"],
+            "STYLES/*": ["./assets/styles/*"],
             "Images/*": ["./assets/images/*"],
             "NODE/*": ["./node_modules/*"]
         }
